requirements.txt
pytest
fastapi
uvicorn
openai
ffmpeg
PyMuPDF
PyPDF2  
python-multipart
pymongo
python-dotenv
Pillow
pytest-mock